Move AssemblyVersionHelper tests to dedicated file and fix unused usings
