+++ tests/Aspire.Cli.Tests/CliBootstrapTests.cs
@@ -6,7 +6,6 @@ using Aspire.Cli.Acquisition;
 using Aspire.Cli.Tests.TestServices;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 
 namespace Aspire.Cli.Tests;
